dia6

+++ dia6.py
@@ -1,3 +1,3 @@
-temperatura_grados = int(input('Ingese la tempretatura en grados centigrados: '))
+temperatura_grados = float(input('Ingese la tempretatura en grados centigrados: '))
 temperatura_fahrenheit = (temperatura_grados * 9/5) + 32
 print(f'la temperatura en fahrenheit es {temperatura_fahrenheit} F°')
